lease page template add
