--- src/main/java/ru/vldf/sportsportal/controller/router/CommonController.java
@@ -1,31 +1,0 @@
-package ru.vldf.sportsportal.controller.router;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.servlet.ModelAndView;
-import springfox.documentation.annotations.ApiIgnore;
-
-@ApiIgnore
-@Controller
-public class CommonController {
-
-    @GetMapping({"/", "/index"}) // todo: remove temporary redirect
-    public String index() {
-        return "home";
-    }
-
-    @GetMapping("/login")
-    public ModelAndView login() {
-        return new ModelAndView("login");
-    }
-
-    @GetMapping("/registration")
-    public ModelAndView registration() {
-        return new ModelAndView("registration");
-    }
-
-    @GetMapping("/home")
-    public ModelAndView home() {
-        return new ModelAndView("home");
-    }
-}
