+++ src/main/resources/templates/fragments/header.html
@@ -1,0 +1,27 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org"
+      lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>HEADER</title>
+    <script src="../../static/libs/js/jquery-3.3.1.min.js"></script>
+    <script src="../../static/libs/js/bootstrap-4.1.3.min.js"></script>
+    <link href="../../static/libs/css/bootstrap-4.1.3.min.css" rel="stylesheet">
+    <link href="../../static/libs/css/font-awesome-4.7.0.min.css" rel="stylesheet">
+    <link href="../../static/css/lease.css" rel="stylesheet">
+</head>
+<body>
+<header class="header-main" th:fragment="header">
+    <div class="container">
+        <div class="row">
+            <div class="col-md-3">
+                <h3><a href="#">АРЕНДА ПЛОЩАДОК</a></h3>
+                <h6><a href="#"><i class="fa fa-angle-double-left"></i>
+                    <small>НА ГЛАВНУЮ</small>
+                </a></h6>
+            </div>
+        </div>
+    </div>
+</header>
+</body>
+</html>
