+++ src/main/resources/templates/lease/lease.html
@@ -1,0 +1,34 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org"
+      lang="ru">
+<head>
+    <meta charset="UTF-8">
+    <title>Аренда спортивных площадок</title>
+
+    <!-- LIBRARIES -->
+    <script src="../../static/libs/js/jquery-3.3.1.min.js" th:src="@{/libs/js/jquery-3.3.1.min.js}"></script>
+    <script src="../../static/libs/js/bootstrap-4.1.3.min.js" th:src="@{/libs/js/bootstrap-4.1.3.min.js}"></script>
+    <link href="../../static/libs/css/bootstrap-4.1.3.min.css" th:href="@{/libs/css/bootstrap-4.1.3.min.css}"
+          rel="stylesheet">
+    <link href="../../static/libs/css/font-awesome-4.7.0.min.css" th:href="@{/libs/css/font-awesome-4.7.0.min.css}"
+          rel="stylesheet">
+
+    <link href="../../static/css/lease.css" th:href="@{/css/lease.css}"
+          rel="stylesheet">
+</head>
+<body>
+<header class="header-main" th:replace="fragments/header :: header"></header>
+<main class="container container-lease">
+    <div class="row">
+        <div class="col-sm-4">
+
+        </div>
+        <div class="col-sm-8">
+            <ul class="card-deck">
+
+            </ul>
+        </div>
+    </div>
+</main>
+</body>
+</html>
